+++ src/sim/fiber_set.cc
@@ -383,7 +383,7 @@ void FiberSet::allIntersections(Array<FiberSite>& res1, Array<FiberSite>& res2,
     // check what other method gives:
     allIntersections0(res1, res2, max_distance);
     std::clog << "FiberSet::allIntersections0() found " << res1.size() << " intersections\n";
-    for ( unsigned i = 0; i < res1.size(); ++i )
+    for ( size_t i = 0; i < res1.size(); ++i )
         std::clog << res1[i] << " " << res2[i] << "\n";
 #endif
     
@@ -429,7 +429,7 @@ void FiberSet::allIntersections(Array<FiberSite>& res1, Array<FiberSite>& res2,
 #if ( 0 )
     // detailled debug output
     std::clog << "FiberSet::allIntersections()  found " << res1.size() << " intersections \n";
-    for ( unsigned i = 0; i < res1.size(); ++i )
+    for ( size_t i = 0; i < res1.size(); ++i )
         std::clog << res1[i] << " " << res2[i] << "\n";
 #endif
 }
